+++ app/src/main/java/org/edb/main/ServerRequester.java
@@ -6,7 +6,7 @@ public interface ServerRequester {
 
     public void requestAvailableExternalServices();
     public void requestUserExternalServices();
-    public void requestExternalServiceDetails();
+    public void requestExternalServiceDetails(int externalIdx);
     public void requestLogin(String id, String pw);
 
 }
