+++ app/src/main/java/org/edb/main/network/RestAPIRequester.java
@@ -5,6 +5,7 @@ import javafx.application.Platform;
 import org.edb.main.*;
 import org.edb.main.UI.UserExternalServiceListController;
 import org.edb.main.network.get.getAvailableExternalServiceResponse;
+import org.edb.main.network.get.getExternalServiceDetailListResponse;
 import org.edb.main.network.get.getExternalServiceListResponse;
 import org.edb.main.network.post.postLoginResponse;
 import retrofit2.Call;
@@ -66,8 +67,6 @@ public class RestAPIRequester  implements ServerRequester {
 
     }
 
-
-
     @Override
     public void requestUserExternalServices() {
         String token=getToken();
@@ -97,8 +96,40 @@ public class RestAPIRequester  implements ServerRequester {
     }
 
     @Override
-    public void requestExternalServiceDetails() {
+    public void requestExternalServiceDetails(int externalIdx) {
 
+        String token=getToken();
+
+        //선택한 외부서비스의 목표달성테이블 데이터받기
+        //요청값 externalIdx,token //응답값  external_service_detail_idx(목표달성idx),name (목표달성이름),if_achieve(달성여부 1 = 목표 달성 , 0 = 목표달성x)
+        //@GET("external/detail/{externalIdx}")
+
+        //데이터 받기 성공시 테이블과 체크박스 보여주기
+        Call<getExternalServiceDetailListResponse> getExternalServiceDetailListResponseCall =
+                RestApiConnector.getExternalServiceNetworkService().getExternalServiceDetailListAPI(externalIdx, token);
+
+        getExternalServiceDetailListResponseCall.enqueue(new Callback<getExternalServiceDetailListResponse>() {
+
+            @Override
+            public void onResponse(Call<getExternalServiceDetailListResponse> call, Response<getExternalServiceDetailListResponse> response) {
+                try {
+                    if (response.isSuccessful()) {
+                        //데이터 받기 성공시 테이블과 체크박스 보여주기
+                        serverResponseHandler.handleExternalServiceDetailsResponse(externalIdx,response.body().getData());
+                    }
+
+                } catch (Exception e) {
+                    e.printStackTrace();
+                }
+            }
+
+            @Override
+            public void onFailure(Call<getExternalServiceDetailListResponse> call, Throwable throwable) {
+                System.out.print("error\n");
+                System.out.println(throwable);
+            }
+
+        });
     }
 
     @Override
